requirements.txt
functions-framework==3.*
google-cloud-logging==3.5.0
requests>=2.28.0
google-cloud-pubsub>=2.13.0
redis>=5.0.0
google-cloud-secret-manager
nltk
langchain-google-vertexai
langchain
python-igraph
pytest
flask
